Remade player pos to float so it can be controlled, created an update function for players, locally and over server

+++ bomberman/player/player.c
@@ -4,7 +4,7 @@
 
 #include "player.h"
 
-Player create_player(int x, int y, int id)
+Player create_player(float x, float y, int id)
 {
     Player ply;
     ply.x = x;
